First pass at nfnl migration (not working)
This uses the old Aniseed macros but manually imported rather than
magically done in the background. I'll then convert the files over one
at a time, updating their test suite to use plenary busted as I go, this
will be a LONG process and I may have to start again at some point if
things get merged into develop.

This is sort of just an experimental branch to see if it's possible to
swap the compiler out without rewriting everything in one go.